requirements.txt
distribute==0.6.34
github3.py==0.7.0
requests==1.2.3
wsgiref==0.1.2
